requirements.txt
requests
smbus2
RPi.bme280
gpiozero
influxdb
pyserial
python-aqi
